src/requirements.txt
openai==1.3.5
python-dotenv==1.0.0
tenacity==8.2.3
tiktoken==0.5.1
Flask==3.0.0